requirements.txt
Flask
tensorflow
keras
numpy
Pillow
gunicorn
flask-cors